Reverted to Windstream
I reverted to the Windstream images that were originally on the pages.
I believe they give the page a certain style that cannot be replaced.
As such, I do not own the material and was in not trying to conflict
with any copyrights. I am just exhibiting work I completed.